move par lock step iter to fn
